Fixed LPTMR, time managing, and screen clearing
Using Low Power Timer to update a system timer every 2ms. Using this to
determine how long to wait at the end of the main loop. Added in
clearing of the screen at the end of a song sequence. Added in
functionality to automatically advance to next effect based on time.